refactor AbstractConnectionFactory
